bug: update with module and fix empty struct
Problem: the current version does not have a go module,
and the build / run will fail with recent versions of
go (cgo) that do not allow creating an empty struct.
This is an attempt to fix that. I will need to test it
to see if it actually works. I am also adding GitHub CI
(to run go test) along with a VSCode developer container
environment that has hwloc ready to go for easier development.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>